src/requirements.txt
backports.zoneinfo==0.2.1;python_version<"3.9"
Django==4.0.5
tzdata==2022.1
psycopg2-binary==2.9.3
asgiref==3.5.2
attrs==21.4.0
django-environ==0.9.0
django-jet==1.0.8
django-jet-reboot==1.3.1
feedparser==6.0.10
iniconfig==1.1.1
packaging==21.3
Pillow==9.1.1
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
pyparsing==3.0.9
pytest==7.1.2
pytest-django==4.5.2
python-dotenv==0.20.0
sgmllib3k==1.0.0
sqlparse==0.4.2
tomli==2.0.1
factory-boy==3.2.1
pytest-factoryboy==2.5.0
django-colorfield==0.7.2
django-ckeditor==6.4.2
django_render_partial==0.4
django-crispy-forms==1.14.0
django-guardian==2.4.0
authlib==1.0.1
requests==2.28.1
# django-allauth==0.51.0
Celery==5.2.7
redis==4.3.4
django-rosetta==0.9.8
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-spectacular==0.24.0
django-modeltranslation==0.18.4
django-six==1.0.4
django-cors-headers==3.13.0
django-celery-beat==2.4.0
django-templated-email==3.0.0
django-utils-six==2.0
django-social-share==2.3.0
mongoengine==0.24.2
pymongo==4.3.2
djongo==1.3.6
pylint-mongoengine==0.4.0
django-redis==5.2.0
django-rest-knox==4.2.0
